// src/types/inventory.d.ts
export interface Variant {
    id: string;
    sku: string;
    attributes: Record<string, string>;
    quantity: number;
    sellingPrice: number;
}
export interface Attribute {
    name: string;
    values: string[];
}
export interface Product {
    id: string;
    name: string;
    sku: string;
    barcode?: string;
    description?: string;
    categoryId: string;
    categoryName?: string;
    costPrice: number;
    sellingPrice: number;
    quantity: number;
    supplierInfo?: string;
    imageUrls?: string[];
    isVariable?: boolean;
    variants?: Variant[];
    options?: Attribute[];
    lowStockThreshold: number;
    createdAt: any;
    updatedAt: any;
}
export interface Category {
    id: string;
    name: string;
    parentId: string | null;
}
export interface Supplier {
    id: string;
    name: string;
    contactName?: string;
    email?: string;
    phone?: string;
    address?: string;
    createdAt: any;
}
export type OrderStatus = "Pending" | "Received" | "Cancelled";
export interface OrderItem {
    productId: string;
    productName: string;
    sku: string;
    quantity: number;
    costPrice: number;
    imageUrl?: string;
}
export interface PurchaseOrder {
    id: string;
    orderNumber: string;
    supplierId: string;
    supplierName: string;
    items: OrderItem[];
    totalAmount: number;
    status: OrderStatus;
    notes?: string;
    createdAt: any;
    updatedAt: any;
    receivedAt?: any;
}